pyproject.toml
[project]
name = "recommender-item-based-cf-tg-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "domster704",email = "domster704@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram (>=3.22.0,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "aiohttp-socks (>=0.10.1,<0.11.0)"
]

[tool.poetry]
packages = [{ include = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.1.0,<26.0.0)"
]
